Implement gesture recognition
Currently uses hard coded file and gesture name. There's a problem with
IGesture::get_Name which prevents us from pulling the data. Short term
the answer here is to use the specific gesture database and keep an
indexable array of gesture names available. It'll need properly
investigating in the future.